[D] Compiler change and small code update
Compilers change:
- remove dmd
- add gdc

Code update:
Add ulong type to eliminate any type conversion. Worked slightly faster on my machine (though it is M1 chip). But need to be checked on the original hardware

+++ _parallel/d/main.d
@@ -1,10 +1,10 @@
 private:
-enum int MAX = 440_000_000;
-enum int[10] CACHE = () {
-	int[10] cache;
-	import std.math : pow;
+enum ulong MAX = 440_000_000;
+enum ulong[10] CACHE = () {
+	ulong[10] cache;
+    import std.math : pow;
 
-	for (int i = 1; i <= 9; ++i) {
+	for (ulong i = 1; i <= 9; ++i) {
 		cache[i] = i.pow(i);
 	}
 
@@ -13,7 +13,7 @@ enum int[10] CACHE = () {
 
 bool isMunchausen(ulong number)
 {
-	int total;
+	ulong total;
 	for (ulong n = number; n > 0; n /= 10)
 	{
 		const auto digit = n % 10;
@@ -34,7 +34,8 @@ void main()
 
 	foreach (n; parallel(iota(MAX))) {
 		if (isMunchausen(n)) {
-			printf("%d\n", n);
+			printf("%ld\n", n);
 		}
 	}
 }
+
